separate remote.go download structures from types (#377)
